+++ app/helpers/hl7_messaging_helper.rb
@@ -1,4 +1,4 @@
-module ChecksumHelper
+module Hl7MessagingHelper
   def render_results_table(subsets:, sites:)
     @subset_request_hash = {}
     @site_request_array = []
@@ -31,11 +31,23 @@ module ChecksumHelper
   def view_checksum_detail(checksum_detail:)
     last = checksum_detail.checksum.nil?
 
-    if last
-      ret = checksum_detail.last_checksum ? checksum_detail.last_checksum : checksum_detail
+    if last && checksum_detail.last_checksum
+        ret = checksum_detail.last_checksum
     else
       last = false
       ret = checksum_detail
+    end
+    [ret, last]
+  end
+
+  def view_discovery_detail(discovery_detail:) #todo cris test this!!
+    last = discovery_detail.hl7_message.nil?
+
+    if last && discovery_detail.last_discovery
+        ret = discovery_detail.last_discovery
+    else
+      last = false
+      ret = discovery_detail
     end
     [ret, last]
   end