+++ app/models/discovery_request.rb
@@ -1,0 +1,17 @@
+require './app/models/HL7Base'
+
+class DiscoveryRequest < ActiveRecord::Base
+  extend HL7RequestBase
+  has_many :discovery_details, :dependent => :destroy
+
+  alias_method(:details, :discovery_details)
+
+  def self.last_discovery_detail(domain, subset, site_id)
+    sql = sql_template(domain, subset, site_id, 'DISCOVERY', 'hl7_message')
+    DiscoveryRequest.connection.select_all(sql).first['last_detail_id']
+  end
+
+end
+=begin
+load('./app/models/discovery_request.rb')
+=end