+++ config/environments/development.rb
@@ -6,6 +6,8 @@ Rails.application.configure do
   # since you don't have to restart the web server when you make code changes.
   config.cache_classes = false
 
+  config.action_controller.allow_forgery_protection = false
+
   # Do not eager load code on boot.
   config.eager_load = false
 
@@ -34,6 +36,8 @@ Rails.application.configure do
   # Checks for improperly declared sprockets dependencies.
   # Raises helpful error messages.
   config.assets.raise_runtime_errors = true
+
+  config.active_support.test_order = :sorted
 
   # Raises error for missing translations
   # config.action_view.raise_on_missing_translations = true