+++ test/integration/hl7_messaging_test.rb
@@ -2,26 +2,35 @@ $testing = true
 require 'test_helper'
 require './config/initializers/01_prisme_init'
 
+
 class Hl7MessagingTest < ActionDispatch::IntegrationTest
 
-  def request_checksum(observing = false)
-    @cr = ChecksumRequest.new
-    @cr.username = 'Cris'
-    @cr.domain = 'Allergy'
+  def setup_test_user
+    @user = User.new
+    @user.email = 'test@tester.com'
+    @user.password = @user.email
+    @user.add_role(Roles::SUPER_USER)
+    @user.save
+  end
+
+  def request(klass, observing = false, hl7_method, keyword)
+    @record = klass.send(:new)
+    @record.username = 'Cris'
+    @record.domain = 'Allergy'
     subset = 'Reactants'
     site_ids = []
     ['443', '442'].each do |site_id|
       site_ids << {va_site_id: site_id}
     end
-    cd_array = @cr.checksum_details.build site_ids
+    cd_array = @record.details.build site_ids
     cd_array.each do |cd|
       cd.subset = subset
       cd.status = JIsaacLibrary::Task::NOT_STARTED
     end
-    @cr.save!
-    task_array = HL7Messaging.get_check_sum_task(checksum_detail_array: @cr.checksum_details.to_a)
+    @record.save!
+    task_array = HL7Messaging.send(hl7_method, {keyword => @record.details.to_a})
     task_to_detail_map = {}
-    details = @cr.checksum_details.to_a
+    details = @record.details.to_a
     # puts "Task array (#{task_array.length}) = #{task_array.toString}"
     # puts "details (#{details.length})= #{details.inspect}"
     task_array.zip(details) do |task, detail|
@@ -29,12 +38,12 @@ class Hl7MessagingTest < ActionDispatch::IntegrationTest
     end
     if observing
       task_array.each do |task|
-        @observer = HL7Messaging::HL7ChecksumObserver.new(task_to_detail_map[task], task)
+        @observer = HL7Messaging::HL7ChecksumDiscoveryObserver.new(task_to_detail_map[task], task)
         runnable = -> do
           task.stateProperty.addListener(@observer)
+          @observer.initial_state_check
         end
         javafx.application.Platform.runLater(runnable)
-        @observer.initial_state_check
       end
     end
     #HL7Messaging.start_hl7_task(task: task)
@@ -43,36 +52,13 @@ class Hl7MessagingTest < ActionDispatch::IntegrationTest
     end
   end
 
-  #
-  # def request_discovery(observing = false)
-  #   @cr = ChecksumRequest.new
-  #   @cr.username = 'Cris'
-  #   @cr.subset_group = 'Allergy'
-  #   @cr.status = JIsaacLibrary::Task::NOT_STARTED
-  #   subset = 'Reactants'
-  #   site_ids = []
-  #   ['443','442'].each do |site_id|
-  #     site_ids << {va_site_id: site_id}
-  #   end
-  #   cd_array = @cr.checksum_details.build site_ids
-  #   cd_array.each do |cd|
-  #     cd.subset = subset
-  #   end
-  #   @cr.save!
-  #   #the task array order aligns with the order of checksum details
-  #   details = @cr.checksum_details.to_a
-  #   task_array = HL7Messaging.get_discovery_task(discovery_detail_array: details)
-  #   task_to_detail_map = {}
-  #   task_array.zip(details) do |task,detail| task_to_detail_map[task] = detail
-  #   if observing
-  #     task_array.each do |task|
-  #       @observer = HL7Messaging::HL7ChecksumObserver.new @cr
-  #       task.stateProperty.addListener(@observer)
-  #     end
-  #   end
-  #   HL7Messaging.start_hl7_task(task: task)
-  #   HL7Messaging.fetch_result(task: task) #blocking call
-  # end
+  def request_checksum(observing = false)
+    request(ChecksumRequest, observing, :get_check_sum_task, :checksum_detail_array)
+  end
+
+  def request_discovery(observing = false)
+    request(DiscoveryRequest, observing, :get_discovery_task, :discovery_detail_array)
+  end
 
   setup do
     @subset_string='some_string'
@@ -87,18 +73,19 @@ class Hl7MessagingTest < ActionDispatch::IntegrationTest
 
   end
 
-  # test 'discovery' do
-  #   result = request_discovery
-  #   if(result.is_a? java.lang.Exception)
-  #     puts "#{result}"
-  #     puts result.backtrace.join("\n")
-  #   end
-  #   #result = 'fizzle' #to force a failure
-  #   assert(result.eql?('done'), 'Expected a result of done, received a result of ' + result.to_s)
-  # end
 
   test 'check_sum' do
     result = request_checksum
+    if (result.is_a? java.lang.Exception)
+      puts "#{result}"
+      puts result.backtrace.join("\n")
+    end
+    #result = 'fizzle' #to force a failure
+    assert(result.eql?(nil), 'Expected a result of nil, received a result of ' + result.to_s)
+  end
+
+  test 'discovery' do
+    result = request_discovery
     if (result.is_a? java.lang.Exception)
       puts "#{result}"
       puts result.backtrace.join("\n")
@@ -118,6 +105,11 @@ class Hl7MessagingTest < ActionDispatch::IntegrationTest
     assert(@observer.new_value == JIsaacLibrary::Task::SUCCEEDED, "The final state of our checksum task was (between the arrows)-->#{@observer.new_value}<--, the old value is -->#{@observer.old_value}<--")
   end
 
+  test 'check_discovery_observer' do
+    request_discovery(true)
+    assert(@observer.new_value == JIsaacLibrary::Task::SUCCEEDED, "The final state of our checksum task was (between the arrows)-->#{@observer.new_value}<--, the old value is -->#{@observer.old_value}<--")
+  end
+
   test 'application_property_url' do
     assert(@app_prop.getInterfaceEngineURL.is_a? String)
   end
@@ -133,6 +125,20 @@ class Hl7MessagingTest < ActionDispatch::IntegrationTest
   test 'HL7_still_running' do
     running = HL7Messaging.running?
     assert(running, 'The HL7 engine is not running!!! (It should be)')
+  end
+
+  #todo Greg fix.
+  #Controller test cases
+  test 'can_request_checksum' do
+    now = Time.now
+    setup_test_user
+    domains = %q([{"id":"Allergy","text":"Allergy","subsets":[{"id":"j2_3","text":"Reactions","icon":true,"li_attr":{"id":"j2_3"},"a_attr":{"href":"#","id":"j2_3_anchor"},"state":{"loaded":true,"opened":false,"selected":true,"disabled":false},"data":{},"parent":"Allergy"}]}])
+    sites = %q([{"id":950,"text":"STLVETSDEV"}])
+    post user_session_path, 'user[email]' => @user.email, 'user[password]' => @user.email
+    post hl7_messaging_results_table_path, nav_type: 'checksum', subset_selections: domains, site_selections: sites
+    requests = ChecksumRequest.where('created_at > ?', now).to_a
+    assert(requests.length == 1, "Too many checksum requests were found!  Expected 1, got #{requests.length}")
+    assert(requests.first.checksum_details.length == 1, "Too many checksum details were found!  Expected 1, got #{requests.first.checksum_details.length}")
   end
 
 end